New way of getting the database update
Now we drop the database fully and get the whole database up to date.
Also, removed the id checkinf to search for duplicated entries as we are
starting new each time. It works with a proxy now. Removed extra fields
in the database. Minimal login. Now requires user and password as
fullmoviesonvimeo subreddit is private.